added atests and utests for shortdoc fix
