+++ atest/robot/libdoc/html_output_from_xmlhtml.robot
@@ -8,6 +8,9 @@ Run Libdoc to XML:HTML and to HTML and Parse Model
     Run Libdoc And Set Output    --format XML:HTML ${TESTDATADIR}/module.py ${OUTXML} 
     Run Libdoc And Parse Model From HTML    ${OUTXML}
 
+*** Comments ***
+This test suite will be changed with one of the next Tasks to contain a check for the roundtrip from library into XML then from XML to html.
+
 *** Test Cases ***
 Name
     ${MODEL}[name]          module
